1238: begin rename of CaseExternal to ElectronicPetition
