+++ shared/src/business/entities/cases/ElectronicPetition.ts
